Cast result to integer for Python 3 compatibility

+++ python/examples/show_png.py
@@ -20,8 +20,8 @@ unicorn.brightness(0.4)
 
 img = Image.open('lofi.png')
 
-for o_x in range(img.size[0]/8):
-	for o_y in range(img.size[1]/8):
+for o_x in range(int(img.size[0]/8)):
+	for o_y in range(int(img.size[1]/8)):
 
 		for x in range(8):
 			for y in range(8):
